io_redir fixed

+++ src/execute/execute_pipe.c
@@ -6,13 +6,14 @@
 /*   By: kshim <kshim@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/30 11:01:23 by mikim3            #+#    #+#             */
-/*   Updated: 2023/01/04 17:01:09 by kshim            ###   ########.fr       */
+/*   Updated: 2023/01/04 18:06:22 by kshim            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../include/ft_minishell.h"
 #include <fcntl.h>
 
+#include "../../include/ft_wrapper_functions.h"
 
 // redirection에 실패하는 케이스 존재할 수 있음
 // redir_fd 할 때 대상 fd가 available한지 확인하는 부분 넣으면 좋을 것 같음
@@ -27,12 +28,14 @@ int	execute_redir(t_tree_node *node, t_pipe *m_pipe)
 	redir_fd = -1;
 	redir = ((t_tree_redir *)node->content)->redir;
 	file_name = ((t_tree_redir *)node->content)->file_name;
+	printf("io_redir - %s\n", redir);
 	if (ft_isdigit(*redir) == BOOL_TRUE)
 	{
 		redir_fd = ft_atoi(redir);
+		printf("redir_fd - %d\n", redir_fd);
 		if (redir_fd < 0)
 			return (FT_ERROR);
-		while (ft_isdigit(*redir) == BOOL_FALSE)
+		while (ft_isdigit(*redir) == BOOL_TRUE)
 			redir++;
 	}
 	if (ft_strcmp(redir, "<") == FT_SUCCESS)
@@ -63,15 +66,16 @@ int	ft_redir_infile(char *file_name, t_pipe *m_pipe, int redir_fd)
 {
 	int	file_fd;
 
-	if (access(file_name, F_OK | R_OK) == 0)
-	{
-		// pipex와 동작 다르니 주의 (파일 redir 전에 redir이미 했을 수가 있음)
-		file_fd= ft_open(file_name, O_RDWR);
-	}
+	if (redir_fd != -1 && ft_fstat(redir_fd, 0) == -1)
+		return (FT_ERROR);
+	if (ft_access(file_name, F_OK | R_OK) == FT_ERROR)
+		return (FT_ERROR);
 	else
-		return (FT_ERROR);
-	if (file_fd== -1)
-		return (FT_ERROR);
+	{
+		file_fd= ft_open(file_name, O_RDWR, 0);
+		if (file_fd== -1)
+			return (FT_ERROR);
+	}
 	if (m_pipe->mnsh_builtin == BOOL_TRUE)
 	{
 		if (redir_fd == -1)
@@ -105,7 +109,6 @@ int	ft_redir_infile(char *file_name, t_pipe *m_pipe, int redir_fd)
 				return (FT_ERROR);
 		}
 	}
-
 	return (FT_SUCCESS);
 }
 
@@ -113,6 +116,8 @@ int	ft_redir_outfile(char *file_name, t_pipe *m_pipe, int redir_fd)
 {
 	int	file_fd;
 
+	if (redir_fd != -1 && ft_fstat(redir_fd, 0) == -1)
+		return (FT_ERROR);
 	file_fd= ft_open(file_name, O_CREAT | O_TRUNC | O_RDWR, 0666);
 	if (file_fd== -1)
 		return (FT_ERROR);
@@ -156,6 +161,8 @@ int	ft_redir_append(char *file_name, t_pipe *m_pipe, int redir_fd)
 {
 	int	file_fd;
 
+	if (redir_fd != -1 && ft_fstat(redir_fd, 0) == -1)
+		return (FT_ERROR);
 	file_fd= ft_open(file_name, O_CREAT | O_APPEND | O_RDWR, 0666);
 	if (file_fd== -1)
 		return (FT_ERROR);
@@ -199,7 +206,9 @@ int	ft_redir_here_doc(t_pipe *m_pipe, int redir_fd)
 {
 	int	file_fd;
 
-	file_fd = ft_open("/tmp/.mnsh_here_doc.tmp", O_RDONLY);
+	if (redir_fd != -1 && ft_fstat(redir_fd, 0) == -1)
+		return (FT_ERROR);
+	file_fd = ft_open("/tmp/.mnsh_here_doc.tmp", O_RDONLY, 0);
 	if (file_fd== -1)
 		return (FT_ERROR);
 	if (unlink("/tmp/.mnsh_here_doc.tmp") < 0)
